Remove useless css
