--- src/component/Content/Content.jsx
@@ -1,7 +1,0 @@
-const Content = (props) => {
-    return ( 
-        <></>
-     );
-}
- 
-export default Content;
