--- src/component/Content/index.js
@@ -1,4 +1,0 @@
-import HomeCard from './CardContent/HomeCard';
-
- 
-export { HomeCard };
